Fix import

+++ training/training_mistral.py
@@ -23,7 +23,7 @@ dataset["chat_sample"][0]
 
 model = AutoModelForCausalLM.from_pretrained(
     base_model,
-    torch_dtype=torch.tf32,
+    torch_dtype=torch.float16,
     device_map={"": 0},
     use_flash_attention_2=True
 )
